Implemented spring security

+++ src/main/java/com/exchange/bank/dao/repositories/ConversionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConversionRepository extends JpaRepository<Conversion, Integer> {
+public interface ConversionRepository extends JpaRepository<Conversion, Long> {
 }
